# api/openapi.yaml
components:
    schemas:
...
        AccessLogsConfig:
            properties:
                enabled:
                    type: boolean
                format:
                    type: string
                path:
                    type: string
            required:
                - enabled
            type: object
...
        BlueGreenConfig:
            properties:
                ActiveVersion:
                    type: string
                AutoPromote:
                    type: boolean
                StandbyVersion:
                    type: string
            required:
                - ActiveVersion
                - StandbyVersion
                - AutoPromote
            type: object
        BootstrapConfig:
            properties:
                adminPort:
                    format: int32
                    minimum: 0
                    type: integer
                staticResources:
                    additionalProperties: true
                    type: object
                xdsClusterName:
                    type: string
            type: object
        CanaryConfig:
            properties:
                BaselineVersion:
                    type: string
                CanaryVersion:
                    type: string
                CanaryWeight:
                    format: int64
                    type: integer
                MatchCriteria:
                    $ref: '#/components/schemas/MatchCriteria'
            required:
                - BaselineVersion
                - CanaryVersion
                - CanaryWeight
            type: object
        Condition:
            properties:
                lastTransitionTime:
                    format: date-time
                    type: string
                message:
                    type: string
                reason:
                    type: string
                status:
                    type: string
                type:
                    type: string
            required:
                - type
                - status
                - lastTransitionTime
            type: object
        ConflictPolicy:
            enum:
                - strict
                - warn
                - takeover
            type: string
        DeleteResponse:
            properties:
                message:
                    type: string
            required:
                - message
            type: object
...
        DeploymentStrategyConfig:
            properties:
                blue_green:
                    $ref: '#/components/schemas/BlueGreenConfig'
                canary:
                    $ref: '#/components/schemas/CanaryConfig'
                type:
                    type: string
            required:
                - type
            type: object
...
        ErrorResponse:
            properties:
                code:
                    format: int64
                    type: integer
                details:
                    additionalProperties:
                        type: string
                    type: object
                error:
                    type: string
            required:
                - error
                - code
            type: object
...
        GatewayProfilePutRequest:
            properties:
                metadata:
                    properties:
                        conflictPolicy:
                            $ref: '#/components/schemas/ConflictPolicy'
                        labels:
                            additionalProperties:
                                type: string
                            type: object
                    type: object
                spec:
                    $ref: '#/components/schemas/GatewayProfileSpec'
            required:
                - spec
            type: object
        GatewayProfileResponse:
            properties:
                kind:
                    $ref: '#/components/schemas/ResourceKind'
                metadata:
                    $ref: '#/components/schemas/ResourceMeta'
                spec:
                    $ref: '#/components/schemas/GatewayProfileSpec'
                status:
                    $ref: '#/components/schemas/GatewayProfileStatus'
            required:
                - kind
                - metadata
                - spec
            type: object
        GatewayProfileSpec:
            properties:
                allowedPolicies:
                    items:
                        type: string
                    type: array
                bootstrap:
                    $ref: '#/components/schemas/BootstrapConfig'
                defaultHttpFilters:
                    items:
                        $ref: '#/components/schemas/HTTPFilter'
                    type: array
                defaultStrategy:
                    $ref: '#/components/schemas/StrategyConfig'
                description:
                    type: string
                displayName:
                    type: string
                envoyImage:
                    type: string
                listenerPresets:
                    items:
                        $ref: '#/components/schemas/ListenerPreset'
                    type: array
                profileType:
                    type: string
            required:
                - displayName
                - profileType
            type: object
        GatewayProfileStatus:
            properties:
                conditions:
                    items:
                        $ref: '#/components/schemas/Condition'
                    type: array
                gatewayCount:
                    format: int64
                    type: integer
                phase:
                    type: string
            required:
                - phase
            type: object
        GatewayPutRequest:
            properties:
                metadata:
                    properties:
                        conflictPolicy:
                            $ref: '#/components/schemas/ConflictPolicy'
                        labels:
                            additionalProperties:
                                type: string
                            type: object
                    type: object
                spec:
                    $ref: '#/components/schemas/GatewaySpec'
            required:
                - spec
            type: object
        GatewayResponse:
            properties:
                kind:
                    $ref: '#/components/schemas/ResourceKind'
                metadata:
                    $ref: '#/components/schemas/ResourceMeta'
                spec:
                    $ref: '#/components/schemas/GatewaySpec'
                status:
                    $ref: '#/components/schemas/GatewayStatus'
            required:
                - kind
                - metadata
                - spec
            type: object
        GatewaySpec:
            properties:
                defaults:
                    $ref: '#/components/schemas/StrategyConfig'
                nodeId:
                    type: string
                profileRef:
                    type: string
            required:
                - nodeId
            type: object
        GatewayStatus:
            properties:
                conditions:
                    items:
                        $ref: '#/components/schemas/Condition'
                    type: array
                phase:
                    type: string
            required:
                - phase
            type: object
        HTTPFilter:
            properties:
                config:
                    additionalProperties: true
                    type: object
                name:
                    type: string
            required:
                - name
                - config
            type: object
        HealthCheckConfig:
            properties:
                enabled:
                    type: boolean
                expected_status:
                    format: int32
                    minimum: 0
                    type: integer
                healthy_count:
                    format: int32
                    minimum: 0
                    type: integer
                interval:
                    type: string
                path:
                    type: string
                timeout:
                    type: string
                unhealthy_count:
                    format: int32
                    minimum: 0
                    type: integer
            required:
                - enabled
            type: object
...
        ListenerListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/ListenerResponse'
                    type: array
                kind:
                    type: string
                total:
                    type: integer
            required:
                - kind
                - items
                - total
            type: object
        ListenerPreset:
            properties:
                address:
                    type: string
                http2:
                    type: boolean
                name:
                    type: string
                port:
                    format: int32
                    minimum: 0
                    type: integer
                tls:
                    $ref: '#/components/schemas/TLSConfig'
            required:
                - name
                - port
            type: object
...
        ListenerResponse:
            properties:
                kind:
                    $ref: '#/components/schemas/ResourceKind'
                metadata:
                    $ref: '#/components/schemas/ResourceMeta'
                spec:
                    $ref: '#/components/schemas/ListenerSpec'
                status:
                    $ref: '#/components/schemas/ListenerStatus'
            required:
                - kind
                - metadata
                - spec
            type: object
        ListenerSpec:
            properties:
                address:
                    type: string
                gatewayRef:
                    type: string
                http2:
                    type: boolean
                port:
                    format: int32
                    minimum: 0
                    type: integer
                tls:
                    $ref: '#/components/schemas/TLSConfig'
            required:
                - gatewayRef
                - port
            type: object
        ListenerStatus:
            properties:
                conditions:
                    items:
                        $ref: '#/components/schemas/Condition'
                    type: array
                phase:
                    type: string
            required:
                - phase
            type: object
        LoadBalancingStrategyConfig:
            properties:
                choice_count:
                    format: int32
                    minimum: 0
                    type: integer
                cookie_name:
                    type: string
                hash_on:
                    type: string
                header_name:
                    type: string
                health_check:
                    $ref: '#/components/schemas/HealthCheckConfig'
                type:
                    type: string
            required:
                - type
            type: object
...
        MetricsConfig:
            properties:
                enabled:
                    type: boolean
                path:
                    type: string
                port:
                    format: int32
                    minimum: 0
                    type: integer
                provider:
                    type: string
            required:
                - enabled
            type: object
        ObservabilityStrategyConfig:
            properties:
                access_logs:
                    $ref: '#/components/schemas/AccessLogsConfig'
                metrics:
                    $ref: '#/components/schemas/MetricsConfig'
                tracing:
                    $ref: '#/components/schemas/TracingConfig'
            type: object
...
        RateLimitStrategyConfig:
            properties:
                burst_size:
                    format: int32
                    minimum: 0
                    type: integer
                claim_name:
                    type: string
                cookie_name:
                    type: string
                external_service:
                    type: string
                header_name:
                    type: string
                identify_by:
                    type: string
                requests_per_minute:
                    format: int32
                    minimum: 0
                    type: integer
                type:
                    type: string
            required:
                - type
            type: object
        ResourceKind:
            enum:
                - Gateway
                - GatewayProfile
                - Listener
                - Environment
                - API
                - Deployment
            type: string
        ResourceMeta:
            properties:
                conflictPolicy:
                    $ref: '#/components/schemas/ConflictPolicy'
                createdAt:
                    format: date-time
                    type: string
                kind:
                    $ref: '#/components/schemas/ResourceKind'
                labels:
                    additionalProperties:
                        type: string
                    type: object
                managedBy:
                    type: string
                name:
                    type: string
                revision:
                    format: int64
                    type: integer
                updatedAt:
                    format: date-time
                    type: string
            required:
                - kind
                - name
                - revision
                - createdAt
                - updatedAt
            type: object
        RetryStrategyConfig:
            properties:
                budget_percent:
                    format: double
                    type: number
                max_retries:
                    format: int32
                    minimum: 0
                    type: integer
                per_try_timeout:
                    type: string
                retriable_status_codes:
                    items:
                        format: int32
                        minimum: 0
                        type: integer
                    type: array
                retry_on:
                    type: string
                type:
                    type: string
            required:
                - type
            type: object
        RouteMatchStrategyConfig:
            properties:
                case_sensitive:
                    type: boolean
                type:
                    type: string
                version_header:
                    type: string
            required:
                - type
            type: object
...
        StrategyConfig:
            properties:
                deployment:
                    $ref: '#/components/schemas/DeploymentStrategyConfig'
                load_balancing:
                    $ref: '#/components/schemas/LoadBalancingStrategyConfig'
                observability:
                    $ref: '#/components/schemas/ObservabilityStrategyConfig'
                rate_limiting:
                    $ref: '#/components/schemas/RateLimitStrategyConfig'
                retry:
                    $ref: '#/components/schemas/RetryStrategyConfig'
                route_matching:
                    $ref: '#/components/schemas/RouteMatchStrategyConfig'
            type: object
        TLSConfig:
            properties:
                caPath:
                    type: string
                certPath:
                    type: string
                cipherSuites:
                    items:
                        type: string
                    type: array
                keyPath:
                    type: string
                minVersion:
                    type: string
                requireClientCert:
                    type: boolean
            required:
                - certPath
                - keyPath
            type: object
        TracingConfig:
            properties:
                enabled:
                    type: boolean
                endpoint:
                    type: string
                provider:
                    type: string
                sampling_rate:
                    format: double
                    type: number
            required:
                - enabled
            type: object
...
info:
    description: Declarative Envoy xDS control plane with reconciliation-based architecture. Resources (Gateways, Listeners, Environments, APIs, Deployments) are declared via this REST API and stored in a desired-state store.
    license:
        name: Apache 2.0
    title: FlowC Control Plane API
    version: 3.0.0
openapi: 3.0.3
paths:
...
    /api/v1/gatewayprofiles/{name}:
        delete:
            operationId: deleteGatewayProfile
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                    description: Resource deleted
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Delete a GatewayProfile
            tags:
                - GatewayProfiles
        get:
            operationId: getGatewayProfile
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayProfileResponse'
                    description: The GatewayProfile resource
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Get a GatewayProfile
            tags:
                - GatewayProfiles
        parameters:
            - description: Resource name
              in: path
              name: name
              required: true
              schema:
                type: string
        put:
            operationId: putGatewayProfile
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
                - description: Identifier of the managing entity for ownership tracking
                  in: header
                  name: X-Managed-By
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GatewayProfilePutRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayProfileResponse'
                    description: Resource updated
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayProfileResponse'
                    description: Resource created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Create or update a GatewayProfile
            tags:
                - GatewayProfiles
...
    /api/v1/gateways/{name}:
        delete:
            operationId: deleteGateway
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DeleteResponse'
                    description: Resource deleted
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Delete a Gateway
            tags:
                - Gateways
        get:
            operationId: getGateway
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayResponse'
                    description: The Gateway resource
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Get a Gateway
            tags:
                - Gateways
        parameters:
            - description: Resource name
              in: path
              name: name
              required: true
              schema:
                type: string
        put:
            operationId: putGateway
            parameters:
                - description: Resource revision for optimistic concurrency control
                  in: header
                  name: If-Match
                  schema:
                    format: int64
                    type: integer
                - description: Identifier of the managing entity for ownership tracking
                  in: header
                  name: X-Managed-By
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GatewayPutRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayResponse'
                    description: Resource updated
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GatewayResponse'
                    description: Resource created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "409":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: Create or update a Gateway
            tags:
                - Gateways
...
    /api/v1/listeners:
        get:
            operationId: listListeners
            parameters:
                - description: Label filter (comma-separated key=value pairs)
                  in: query
                  name: labels
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListenerListResponse'
                    description: List of listeners
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Error response
            summary: List listeners
            tags:
                - Listeners
...
servers:
    - description: Local development server
      url: http://localhost:8080
...